# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_igeargeek_sales-system-api_internal_models.JobTitleOption:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      deleted_by:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  github_com_igeargeek_sales-system-api_internal_models.LeadSourceOption:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      deleted_by:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  github_com_igeargeek_sales-system-api_internal_models.PipelineStage:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      deleted_by:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      is_lost_stage:
        type: boolean
      is_won_stage:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handlers.jobTitleOptionForm:
    properties:
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
  internal_handlers.leadSourceForm:
    properties:
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
  internal_handlers.pipelineStageForm:
    properties:
      is_active:
        type: boolean
      is_lost_stage:
        type: boolean
      is_won_stage:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    CRM/sales-pipeline API. This spec currently covers the
    Prospect stage, Pipeline stage, and Dashboard summary routes
    as a scaffold — the annotation pattern on those handlers
    (see internal/handlers/prospect_stages.go, pipeline_stages.go,
    dashboard.go) extends to the rest of internal/routes/routes.go
    as they're annotated. Run `go run github.com/swaggo/swag/cmd/swag@v1.16.6
    init -g cmd/api/main.go -o docs --parseDependency --parseInternal` from the
    repo root to regenerate docs/ after editing annotations (the two flags are
    required to resolve gorm.DeletedAt across package boundaries).
  title: Sales System API
  version: "1.0"
paths:
...
  /admin/job-titles:
    get:
      description: Returns every configured Contact job title (active + inactive),
        ordered by name.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.JobTitleOption'
            type: array
      security:
      - BearerAuth: []
      summary: List job titles
      tags:
      - admin/job-titles
    post:
      consumes:
      - application/json
      description: Admin-only.
      parameters:
      - description: Job title fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.jobTitleOptionForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.JobTitleOption'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a job title
      tags:
      - admin/job-titles
  /admin/job-titles/{id}:
    delete:
      description: 'Admin-only. Soft-delete (is_active: false).'
      parameters:
      - description: Job title ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate a job title
      tags:
      - admin/job-titles
    patch:
      consumes:
      - application/json
      description: Admin-only.
      parameters:
      - description: Job title ID
        in: path
        name: id
        required: true
        type: integer
      - description: Job title fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.jobTitleOptionForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.JobTitleOption'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a job title
      tags:
      - admin/job-titles
  /admin/lead-sources:
    get:
      description: Returns every configured lead/deal source (active + inactive),
        ordered by name.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.LeadSourceOption'
            type: array
      security:
      - BearerAuth: []
      summary: List lead sources
      tags:
      - admin/lead-sources
    post:
      consumes:
      - application/json
      description: Admin-only.
      parameters:
      - description: Lead source fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.leadSourceForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.LeadSourceOption'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a lead source
      tags:
      - admin/lead-sources
  /admin/lead-sources/{id}:
    delete:
      description: 'Admin-only. Soft-delete (is_active: false).'
      parameters:
      - description: Lead source ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate a lead source
      tags:
      - admin/lead-sources
    patch:
      consumes:
      - application/json
      description: Admin-only.
      parameters:
      - description: Lead source ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lead source fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.leadSourceForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.LeadSourceOption'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a lead source
      tags:
      - admin/lead-sources
  /admin/pipeline-stages:
    get:
      description: Returns every configured pipeline stage (active + inactive), ordered
        by sort_order.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.PipelineStage'
            type: array
      security:
      - BearerAuth: []
      summary: List pipeline stages
      tags:
      - admin/pipeline-stages
    post:
      consumes:
      - application/json
      description: Admin-only. Setting is_won_stage/is_lost_stage clears that flag
        from every other stage.
      parameters:
      - description: Stage fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.pipelineStageForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.PipelineStage'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a pipeline stage
      tags:
      - admin/pipeline-stages
  /admin/pipeline-stages/{id}:
    delete:
      description: 'Admin-only. Soft-delete (is_active: false) — existing Deals may
        still reference the name.'
      parameters:
      - description: Pipeline stage ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate a pipeline stage
      tags:
      - admin/pipeline-stages
    patch:
      consumes:
      - application/json
      description: Admin-only. Setting is_won_stage/is_lost_stage clears that flag
        from every other stage.
      parameters:
      - description: Pipeline stage ID
        in: path
        name: id
        required: true
        type: integer
      - description: Stage fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.pipelineStageForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_igeargeek_sales-system-api_internal_models.PipelineStage'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a pipeline stage
      tags:
      - admin/pipeline-stages
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT, e.g. "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"